+++ src/lib/barrier/key_types.cpp
@@ -85,6 +85,7 @@ const KeyNameMapEntry kKeyNameMap[] = {
     { "F34",            kKeyF34 },
     { "F35",            kKeyF35 },
     { "Find",            kKeyFind },
+    { "Function",        kKeyFunction },
     { "Help",            kKeyHelp },
     { "Henkan",            kKeyHenkan },
     { "Home",            kKeyHome },